+++ Production/govuk_ios/ViewModels/HomeViewModel.swift
@@ -7,8 +7,7 @@ struct HomeViewModel {
     let analyticsService: AnalyticsServiceInterface
     let configService: AppConfigServiceInterface
     let notificationService: NotificationServiceInterface
-  //  let topicWidgetViewModel: TopicsWidgetViewModel
-    let topicsWidgetViewModelSwiftUI: TopicsWidgetViewModelSwiftUI
+    let topicsWidgetViewModel: TopicsWidgetViewModel
     let localAuthorityAction: () -> Void
     let editLocalAuthorityAction: () -> Void
     let feedbackAction: () -> Void
@@ -145,23 +144,11 @@ struct HomeViewModel {
         return widget
     }
 
-//    @MainActor
-//    private var topicsWidget: WidgetView? {
-//        guard featureEnabled(.topics)
-//        else { return nil }
-//        let content = TopicsWidgetView(
-//            viewModel: topicsWidgetViewModelSwiftUI
-//        )
-//        let widget = WidgetView(decorateView: false)
-//        widget.addContent(content)
-//        return widget
-//    }
-
     private func featureEnabled(_ feature: Feature) -> Bool {
         configService.isFeatureEnabled(key: feature)
     }
 
     func trackECommerce() {
-        topicsWidgetViewModelSwiftUI.trackECommerce()
+        topicsWidgetViewModel.trackECommerce()
     }
 }
