--- Production/govuk_ios/Views/Home/Topics/TopicCardSwiftUi.swift
@@ -1,13 +1,0 @@
-//
-
-import SwiftUI
-
-struct TopicCardSwiftUi: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    TopicCardSwiftUi()
-}
