+++ Tests/govuk_ios/govuk_ios_snapshot_tests/Specs/ViewControllers/EditTopicsViewControllerSnapshotTests.swift
@@ -106,7 +106,6 @@ final class EditTopicsViewControllerSnapshotTests: SnapshotTestCase {
         let viewModel = EditTopicsViewModel(
             topicsService: mockTopicsService,
             analyticsService: MockAnalyticsService(),
-            dismissAction: { }
         )
         let view = EditTopicsView(
             viewModel: viewModel
