+++ Tests/govuk_ios/govuk_ios_unit_tests/Mocks/Builders/MockViewControllerBuilder.swift
@@ -29,7 +29,7 @@ class MockViewControllerBuilder: ViewControllerBuilder {
         _receivedEditLocalAuthorityAction = actions.editLocalAuthorityAction
         _receivedHomeRecentActivityAction = actions.recentActivityAction
         _receivedHomeSearchAction = actions.openSearchAction
-        _receivedTopicWidgetViewModel = dependencies.topicWidgetViewModel
+        _receivedTopicWidgetViewModel = dependencies.topicsWidgetViewModel
         return _stubbedHomeViewController ?? UIViewController()
     }
 
@@ -73,14 +73,6 @@ class MockViewControllerBuilder: ViewControllerBuilder {
         return _stubbedTopicDetailViewController ?? UIViewController()
     }
 
-    var _stubbedEditTopicsViewController: UIViewController?
-    var _receivedDismissAction: (() -> Void)?
-    override func editTopics(analyticsService: any AnalyticsServiceInterface,
-                             topicsService: any TopicsServiceInterface,
-                             dismissAction: @escaping () -> Void) -> UIViewController {
-        _receivedDismissAction = dismissAction
-        return _stubbedEditTopicsViewController ?? UIViewController()
-    }
     var _stubbedLocalAuthorityPostcodeEntryViewController: UIViewController?
     var _receivedLocalAuthorityDismissAction: (() -> Void)?
     var _receivedResolveAmbiguityAction: ((AmbiguousAuthorities, String) -> Void)?
